tasks/main.yml: Allow to specify ssh key options
For now, it's not possible to set ssh key options for a user
key, so add support for it.
So, either the key list entry is a string and there's no key
options or a dict with {key, opt} keys to specify the key and
the key options.
Everything remains compatible with previous configuration, so
as to not break stuff.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>